[ refactor ] : entity 변경사항 적용
