+++ domain/src/main/java/com/ddd/ansayo/domain/model/favorite/FavoriteCourseState.kt
@@ -1,6 +1,6 @@
 package com.ddd.ansayo.domain.model.favorite
 
-import com.ddd.ansayo.domain.model.course.Course
+import com.ddd.ansayo.core_model.course.Course
 
 data class FavoriteCourseState(
     val hasFavorites: Boolean,
